+++ visidata/expr.py
@@ -41,4 +41,4 @@ def inputExpr(self, prompt, *args, **kwargs):
 
 
 Sheet.addCommand('=', 'addcol-expr', 'addColumn(ColumnExpr(inputExpr("new column expr=")), index=cursorColIndex+1)', undo=undoAddCols)
-Sheet.addCommand('g=', 'setcol-expr', 'cursorCol.setValuesFromExpr(selectedRows or rows, inputExpr("set selected="))', undo=undoEditCells)
+Sheet.addCommand('g=', 'setcol-expr', 'cursorCol.setValuesFromExpr(selectedRows, inputExpr("set selected="))', undo=undoEditCells)